Aggregation/requirements.txt
aiokafka==0.10.0
aioredis==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncio==3.4.3
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.0
fastapi-cache==0.1.0
fastapi-cache2==0.2.1
greenlet==3.0.3
h11==0.14.0
hiredis==2.3.2
idna==3.6
packaging==23.2
pendulum==3.0.0
pydantic==2.6.0
pydantic_core==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
redis==5.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
time-machine==2.13.0
typing_extensions==4.9.0
tzdata==2024.1
uvicorn==0.27.0.post1
